Fix unlocks gridview layout
