+++ BF4Intel/src/main/res/layout/grid_item_unlocks.xml
@@ -36,21 +36,11 @@
     android:singleLine="true"
     />
 
-  <TextView
-    android:id="@+id/unlock_subtitle"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:textAppearance="@style/TextAppearance.Holo.Smallest"
-    android:layout_below="@id/unlock_title"
-    android:layout_centerHorizontal="true"
-    android:singleLine="true"
-    />
-
   <ProgressBar
     style="?android:attr/progressBarStyleHorizontal"
     android:id="@+id/unlock_completion"
     android:layout_height="wrap_content"
     android:layout_width="match_parent"
-    android:layout_below="@id/unlock_subtitle"
+    android:layout_below="@id/unlock_title"
     />
 </RelativeLayout>
