+++ core/src/main/java/com/ninetwozero/bf4intel/json/unlocks/KitItemUnlockContainer.java
@@ -2,7 +2,7 @@ package com.ninetwozero.bf4intel.json.unlocks;
 
 import com.google.gson.annotations.SerializedName;
 
-public class KitItemUnlockContainer implements Comparable<KitItemUnlockContainer> {
+public class KitItemUnlockContainer implements Comparable<KitItemUnlockContainer>, UnlockContainer {
     @SerializedName("kitItemUnlock")
     private WeaponUnlock kitItemUnlock;
     @SerializedName("weaponUnlock")
@@ -15,5 +15,10 @@ public class KitItemUnlockContainer implements Comparable<KitItemUnlockContainer
     @Override
     public int compareTo(final KitItemUnlockContainer otherUnlock) {
         return getUnlock().compareTo(otherUnlock.getUnlock());
+    }
+
+    @Override
+    public UnlockCriteria getCriteria() {
+        return getUnlock().getCriteria();
     }
 }
