+++ core/src/main/java/com/ninetwozero/bf4intel/json/unlocks/UnlockContainer.java
@@ -1,0 +1,5 @@
+package com.ninetwozero.bf4intel.json.unlocks;
+
+public interface UnlockContainer {
+    UnlockCriteria getCriteria();
+}
