+++ core/src/main/java/com/ninetwozero/bf4intel/json/unlocks/VehicleUnlock.java
@@ -2,7 +2,7 @@ package com.ninetwozero.bf4intel.json.unlocks;
 
 import com.google.gson.annotations.SerializedName;
 
-public class VehicleUnlock implements Comparable<VehicleUnlock> {
+public class VehicleUnlock implements Comparable<VehicleUnlock>, UnlockContainer {
     @SerializedName("unlockId")
     private String name;
     @SerializedName("guid")
