+++ core/src/main/java/com/ninetwozero/bf4intel/json/unlocks/WeaponUnlockContainer.java
@@ -2,7 +2,7 @@ package com.ninetwozero.bf4intel.json.unlocks;
 
 import com.google.gson.annotations.SerializedName;
 
-public class WeaponUnlockContainer implements Comparable<WeaponUnlockContainer> {
+public class WeaponUnlockContainer implements Comparable<WeaponUnlockContainer>, UnlockContainer {
     @SerializedName("weaponUnlock")
     private WeaponUnlock unlock;
 
@@ -13,5 +13,10 @@ public class WeaponUnlockContainer implements Comparable<WeaponUnlockContainer> 
     @Override
     public int compareTo(final WeaponUnlockContainer otherUnlock) {
         return unlock.compareTo(otherUnlock.getUnlock());
+    }
+
+    @Override
+    public UnlockCriteria getCriteria() {
+        return unlock.getCriteria();
     }
 }
